+++ py-polars/polars/io/_utils.py
@@ -6,11 +6,36 @@ from contextlib import contextmanager
 from io import BytesIO, StringIO
 from pathlib import Path
 from tempfile import NamedTemporaryFile
-from typing import IO, Any, ContextManager, Iterator, cast, overload
+from typing import IO, Any, ContextManager, Iterator, Sequence, cast, overload
 
-from polars._utils.various import normalize_filepath
+from polars._utils.various import is_int_sequence, is_str_sequence, normalize_filepath
 from polars.dependencies import _FSSPEC_AVAILABLE, fsspec
 from polars.exceptions import NoDataError
+
+
+def handle_projection_columns(
+    columns: Sequence[str] | Sequence[int] | str | None,
+) -> tuple[list[int] | None, Sequence[str] | None]:
+    """Disambiguates between columns specified as integers vs. strings."""
+    projection: list[int] | None = None
+    new_columns: Sequence[str] | None = None
+    if columns is not None:
+        if isinstance(columns, str):
+            new_columns = [columns]
+        elif is_int_sequence(columns):
+            projection = list(columns)
+        elif not is_str_sequence(columns):
+            msg = "`columns` arg should contain a list of all integers or all strings values"
+            raise TypeError(msg)
+        else:
+            new_columns = columns
+        if columns and len(set(columns)) != len(columns):
+            msg = f"`columns` arg should only have unique values, got {columns!r}"
+            raise ValueError(msg)
+        if projection and len(set(projection)) != len(projection):
+            msg = f"`columns` arg should only have unique values, got {projection!r}"
+            raise ValueError(msg)
+    return projection, new_columns
 
 
 def _is_glob_pattern(file: str) -> bool:
