+++ py-polars/polars/io/avro.py
@@ -4,8 +4,9 @@ import contextlib
 from pathlib import Path
 from typing import IO, TYPE_CHECKING
 
-from polars._utils.various import handle_projection_columns, normalize_filepath
+from polars._utils.various import normalize_filepath
 from polars._utils.wrap import wrap_df
+from polars.io._utils import handle_projection_columns
 
 with contextlib.suppress(ImportError):  # Module not available when building docs
     from polars.polars import PyDataFrame
